feat!: add a new method for working with code status
BREAKING CHANGE: Use an enum to determine the status of a code instead of a set of flags. This should simplify the logic of the smart contract.

The getCodeStatus method now returns the status as a number from the enum instead of an object.

The isAvailableForMint, isHeld, isBlocked methods have been replaced with one generic method - getCodeStatuses.